:art: rename folder from `responses` to `responseSchemas`
