+++ routes/responseSchemas/hero/HeroDurationsResponse.js
