+++ routes/responseSchemas/hero/HeroItemPopularityResponse.js
