+++ routes/responseSchemas/hero/HeroMatchupsResponse.js
