+++ routes/responseSchemas/hero/HeroObjectResponse.js
