+++ routes/responseSchemas/miscellaneous/LeagueObjectResponse.js
