+++ routes/responseSchemas/miscellaneous/RankingsResponse.js
