+++ routes/responseSchemas/miscellaneous/ScenarioLaneRolesResponse.js
