+++ routes/responseSchemas/miscellaneous/ScenarioMiscResponse.js
