+++ routes/responseSchemas/miscellaneous/SchemaResponse.js
