+++ routes/responseSchemas/miscellaneous/SearchResponse.js
